episode-09/streamlit-ifc-template/requirements.txt
streamlit
ifcopenshell-python
matplotlib
plotly
pandas
xlsxwriter